Macro mirai_annotations::checked_verify_ne

macro_rules! checked_verify_ne {
    ($left:expr, $right:expr) => { ... };
    ($left:expr, $right:expr, $message:literal) => { ... };
    ($left:expr, $right:expr, $($arg:tt)*) => { ... };
}

Equivalent to the standard assert_eq! when used with an unmodified Rust compiler. When compiled with MIRAI, this causes MIRAI to check the condition and emit a diagnostic unless it can prove it to be true.