Type Definition mirai_annotations::TagPropagationSet

pub type TagPropagationSet = u128;

A type used to specify how tag types transfer over operations. The type is an alias of u128. Each bit of the bit vector controls the transfer function for an operation. If a bit is set to one, the corresponding operation will propagate the tag. If a bit is set to zero, the corresponding operation will block the tag.